Moved mkdir and chmod out of podman
